+++ contracts/libraries/LibAppStorage.sol
@@ -6,9 +6,7 @@ import {IntentMeta} from "./IntentTypes.sol";
 
 /// @dev Constant representing the hash of the string "APP_STORAGE_POSITION".
 ///      This is used as a unique identifier for the application storage position.
-bytes32 constant APP_STORAGE_POSITION = keccak256(
-    "diamond.standard.app.storage"
-);
+bytes32 constant APP_STORAGE_POSITION = keccak256("diamond.standard.app.storage");
 
 /// @dev Constant representing the hash of the string "PROVIDER_ROLE".
 ///      This is used as a unique identifier for the provider role within the contract.
@@ -60,7 +58,7 @@ struct LabBase {
 /// @param labId The unique identifier for the laboratory.
 /// @param base The base information of the laboratory, represented by the `LabBase` structure.
 struct Lab {
-    uint labId;
+    uint256 labId;
     LabBase base;
 }
 
@@ -85,50 +83,50 @@ struct Lab {
 /// @param payerInstitution Address of the institution paying for the reservation (zero for wallet payments)
 /// @param collectorInstitution Address of the institution that should receive the payout (zero for direct wallet payouts)
 struct Reservation {
-        uint256 labId;           // Slot 0: 32 bytes
-        address renter;          // Slot 1: 20 bytes
-        uint96 price;            // Slot 1: +12 bytes = 32 bytes total
-        address labProvider;     // Slot 2: 20 bytes
-        uint8 status;            // Slot 2: +1 byte
-        uint32 start;            // Slot 2: +4 bytes
-        uint32 end;              // Slot 2: +4 bytes = 29 bytes total
-        string puc;              // Slot 3: 32 bytes (pointer)
-        uint64 requestPeriodStart; // Slot 4: +8 bytes
-        uint64 requestPeriodDuration; // Slot 4: +8 bytes (0 for wallet reservations)
-        address payerInstitution;   // Slot 5: 20 bytes
-        address collectorInstitution; // Slot 5: +20 bytes (stored in separate slot)
-        uint96 providerShare;      // Slot 6: Provider allocation cached at confirmation
-        uint96 projectTreasuryShare; // Slot 6: +12 bytes
-        uint96 subsidiesShare;     // Slot 7: Allocation for subsidies pool
-        uint96 governanceShare;    // Slot 7: +12 bytes
+    uint256 labId; // Slot 0: 32 bytes
+    address renter; // Slot 1: 20 bytes
+    uint96 price; // Slot 1: +12 bytes = 32 bytes total
+    address labProvider; // Slot 2: 20 bytes
+    uint8 status; // Slot 2: +1 byte
+    uint32 start; // Slot 2: +4 bytes
+    uint32 end; // Slot 2: +4 bytes = 29 bytes total
+    string puc; // Slot 3: 32 bytes (pointer)
+    uint64 requestPeriodStart; // Slot 4: +8 bytes
+    uint64 requestPeriodDuration; // Slot 4: +8 bytes (0 for wallet reservations)
+    address payerInstitution; // Slot 5: 20 bytes
+    address collectorInstitution; // Slot 5: +20 bytes (stored in separate slot)
+    uint96 providerShare; // Slot 6: Provider allocation cached at confirmation
+    uint96 projectTreasuryShare; // Slot 6: +12 bytes
+    uint96 subsidiesShare; // Slot 7: Allocation for subsidies pool
+    uint96 governanceShare; // Slot 7: +12 bytes
 }
 
 struct PayoutCandidate {
-        uint32 end;
-        bytes32 key;
+    uint32 end;
+    bytes32 key;
 }
 
 struct RecentReservationBuffer {
-        bytes32[50] keys;
-        uint32[50] starts;
-        uint8 size;
+    bytes32[50] keys;
+    uint32[50] starts;
+    uint8 size;
 }
 
 struct UpcomingReservationBuffer {
-        bytes32[50] keys;
-        uint32[50] starts;
-        uint8 size;
+    bytes32[50] keys;
+    uint32[50] starts;
+    uint8 size;
 }
 
 struct PastReservationBuffer {
-        bytes32[50] keys;
-        uint32[50] ends;
-        uint8 size;
+    bytes32[50] keys;
+    uint32[50] ends;
+    uint8 size;
 }
 
 struct UserActiveReservation {
-        uint32 start;
-        bytes32 key;
+    uint32 start;
+    bytes32 key;
 }
 
 /// @notice Represents a node in a red-black tree data structure, necessary for the library RivalIntervalTree Node data structure
@@ -139,11 +137,11 @@ struct UserActiveReservation {
 /// @param end The ending value of the interval (the beginning value is stored as the key)
 /// @param red Boolean flag indicating if the node is red (true) or black (false)
 struct Node {
-        uint parent;
-        uint left;
-        uint right;
-        uint end; // begin is implicit as the key
-        bool red;
+    uint256 parent;
+    uint256 left;
+    uint256 right;
+    uint256 end; // begin is implicit as the key
+    bool red;
 }
 
 /// @notice Represents a red-black tree data structure, necessary for the library RivalIntervalTree Tree data structure
@@ -151,8 +149,8 @@ struct Node {
 /// @param root The root hash/value of the Merkle Tree
 /// @param nodes Mapping from uint keys to Node values representing the tree structure
 struct Tree {
-        uint root;
-        mapping(uint => Node) nodes;
+    uint256 root;
+    mapping(uint256 => Node) nodes;
 }
 
 /// @notice Struct representing a provider's staking information
@@ -204,7 +202,7 @@ struct InstitutionalUserSpending {
 
 /// @dev This struct is used to define the storage layout for the application.
 ///       Contains all state variables used across the diamond contract. It contains the following fields:
-/// @notice 
+/// @notice
 /// @custom:storage-layout This struct defines the storage layout for the diamond contract
 /// @custom:member DEFAULT_ADMIN_ROLE Stores the keccak256 hash for admin role
 /// @custom:member labTokenAddress Address of the LAB token contract
@@ -247,32 +245,32 @@ struct AppStorage {
     mapping(bytes32 role => EnumerableSet.AddressSet) roleMembers;
     mapping(address => ProviderBase) providers;
     uint256 labId;
-    mapping(uint => LabBase) labs;
+    mapping(uint256 => LabBase) labs;
 
     mapping(uint256 => Tree) calendars;
-    mapping(bytes32 => Reservation) reservations; 
-    mapping(address => EnumerableSet.Bytes32Set) renters; 
+    mapping(bytes32 => Reservation) reservations;
+    mapping(address => EnumerableSet.Bytes32Set) renters;
     uint256 totalReservationsCount;
-    mapping (uint256 => EnumerableSet.Bytes32Set) reservationKeysByToken;
-    mapping (uint256 => mapping(address => bytes32)) activeReservationByTokenAndUser;
-    mapping (uint256 => mapping(address => uint8)) activeReservationCountByTokenAndUser;
-    mapping (uint256 => mapping(address => EnumerableSet.Bytes32Set)) reservationKeysByTokenAndUser;
-    mapping (uint256 => RecentReservationBuffer) recentReservationsByToken;
-    mapping (uint256 => mapping(address => RecentReservationBuffer)) recentReservationsByTokenAndUser;
-    mapping (uint256 => UpcomingReservationBuffer) upcomingReservationsByToken;
-    mapping (uint256 => mapping(address => UpcomingReservationBuffer)) upcomingReservationsByTokenAndUser;
-    mapping (uint256 => PastReservationBuffer) pastReservationsByToken;
-    mapping (uint256 => mapping(address => PastReservationBuffer)) pastReservationsByTokenAndUser;
-    mapping (uint256 => mapping(address => UserActiveReservation[])) activeReservationHeaps;
-    mapping (bytes32 => bool) activeReservationHeapContains;
-    mapping (uint256 => bool) tokenStatus;
-    mapping (uint256 => uint256) labActiveReservationCount;
-    mapping (address => uint256) providerActiveReservationCount;
-    mapping (uint256 => PayoutCandidate[]) payoutHeaps;
-    mapping (bytes32 => bool) payoutHeapContains;
-    mapping (uint256 => uint256) payoutHeapInvalidCount;
-    
-    mapping (address => ProviderStake) providerStakes;
+    mapping(uint256 => EnumerableSet.Bytes32Set) reservationKeysByToken;
+    mapping(uint256 => mapping(address => bytes32)) activeReservationByTokenAndUser;
+    mapping(uint256 => mapping(address => uint8)) activeReservationCountByTokenAndUser;
+    mapping(uint256 => mapping(address => EnumerableSet.Bytes32Set)) reservationKeysByTokenAndUser;
+    mapping(uint256 => RecentReservationBuffer) recentReservationsByToken;
+    mapping(uint256 => mapping(address => RecentReservationBuffer)) recentReservationsByTokenAndUser;
+    mapping(uint256 => UpcomingReservationBuffer) upcomingReservationsByToken;
+    mapping(uint256 => mapping(address => UpcomingReservationBuffer)) upcomingReservationsByTokenAndUser;
+    mapping(uint256 => PastReservationBuffer) pastReservationsByToken;
+    mapping(uint256 => mapping(address => PastReservationBuffer)) pastReservationsByTokenAndUser;
+    mapping(uint256 => mapping(address => UserActiveReservation[])) activeReservationHeaps;
+    mapping(bytes32 => bool) activeReservationHeapContains;
+    mapping(uint256 => bool) tokenStatus;
+    mapping(uint256 => uint256) labActiveReservationCount;
+    mapping(address => uint256) providerActiveReservationCount;
+    mapping(uint256 => PayoutCandidate[]) payoutHeaps;
+    mapping(bytes32 => bool) payoutHeapContains;
+    mapping(uint256 => uint256) payoutHeapInvalidCount;
+
+    mapping(address => ProviderStake) providerStakes;
 
     mapping(address provider => uint256 balance) institutionalTreasury;
     mapping(address provider => uint256 limit) institutionalUserLimit;
@@ -285,7 +283,7 @@ struct AppStorage {
     mapping(address institution => EnumerableSet.Bytes32Set orgs) institutionSchacHomeOrganizations;
 
     // Revenue split buckets (new split replaces legacy pendingLabPayout/pendingInstitutionalLabPayout)
-    mapping (uint256 => uint256) pendingProviderPayout; // per lab pending amount for provider withdrawals
+    mapping(uint256 => uint256) pendingProviderPayout; // per lab pending amount for provider withdrawals
     uint256 pendingProjectTreasury; // global pending amount for project treasury
     uint256 pendingSubsidies; // global pending amount for student subsidies
     uint256 pendingGovernance; // global pending amount for governance incentives
@@ -333,16 +331,16 @@ struct AppStorage {
 library LibAppStorage {
     /// @notice Base stake required for providers who received initial tokens
     uint256 internal constant BASE_STAKE = 800_000_000; // 800 tokens with 6 decimals
-    
+
     /// @notice Number of labs included in base stake (free labs)
     uint256 internal constant FREE_LABS_COUNT = 10;
-    
+
     /// @notice Additional stake required per lab beyond the free count
     uint256 internal constant STAKE_PER_ADDITIONAL_LAB = 200_000_000; // 200 tokens with 6 decimals
-    
+
     /// @notice Default spending limit for institutional users
     uint256 internal constant DEFAULT_INSTITUTIONAL_USER_LIMIT = 10_000_000; // 10 tokens with 6 decimals
-    
+
     /// @notice Default spending period duration (120 days in seconds)
     uint256 internal constant DEFAULT_SPENDING_PERIOD = 120 days;
